requirements.txt
selenium==4.29.0
webdriver-manager==4.0.2
metapub==0.5.12
requests==2.32.3
python-dotenv==1.0.1
transformers==4.36.2
torch==2.1.2
PyPDF2==3.0.1
python-docx==1.1.0
